% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_sample_IGH_chain_all_metrics_wide_format_dataframe}
\alias{all_sample_IGH_chain_all_metrics_wide_format_dataframe}
\title{a dataframe of wide format}
\format{
A data frame with m rows and n variables:
}
\source{
\url{https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/list}
}
\usage{
all_sample_IGH_chain_all_metrics_wide_format_dataframe
}
\description{
The data frame must be generated from combine_all_sample_repertoire_metrics function
}
\keyword{datasets}
