% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_sample_IGH_chain_all_metrics_wide_format_dataframe_list}
\alias{all_sample_IGH_chain_all_metrics_wide_format_dataframe_list}
\title{a dataframe list to combine rowwise.}
\format{
A list of data frames with m rows and n variables:
}
\source{
\url{https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/list}
}
\usage{
all_sample_IGH_chain_all_metrics_wide_format_dataframe_list
}
\description{
The element data frame must be generated from calculate_IGH_isotype_proportion function
}
\keyword{datasets}
