% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_sample_IGH_chain_individual_metrics_dataframe_list}
\alias{all_sample_IGH_chain_individual_metrics_dataframe_list}
\title{a list of dataframes}
\format{
A list of data frame with m rows and n variables:
}
\source{
\url{https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/list}
}
\usage{
all_sample_IGH_chain_individual_metrics_dataframe_list
}
\description{
The list of data frames generated from get_item_name_x_sample_name_for_each_metric
}
\keyword{datasets}
