% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_chain_name_and_cell_type.R
\name{annotate_chain_name_and_cell_type}
\alias{annotate_chain_name_and_cell_type}
\title{Annotate chain name and cell type}
\usage{
annotate_chain_name_and_cell_type(input_dataframe)
}
\arguments{
\item{input_dataframe}{A dataframe of clonotype of immunarch format}
}
\value{
A dataframe of clonotype of immunarch format
}
\description{
Annotate chain name and cell type
}
\examples{
annotate_chain_name_and_cell_type(nonproductive_CDR3aa_removed_dataframe)

}
