% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_chain_name_and_isotype_name.R
\name{annotate_chain_name_and_isotype_name}
\alias{annotate_chain_name_and_isotype_name}
\title{Annotate chain name and isotype name}
\usage{
annotate_chain_name_and_isotype_name(input_dataframe)
}
\arguments{
\item{input_dataframe}{A dataframe of clonotype of immunarch format}
}
\value{
A dataframe of clonotype of immunarch format
}
\description{
Annotate chain name and isotype name
}
\examples{
annotate_chain_name_and_isotype_name(nonproductive_CDR3aa_removed_dataframe)

}
