% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire_metrics_functions.R
\name{calculate_all_sample_chain_repertoire_metrics}
\alias{calculate_all_sample_chain_repertoire_metrics}
\title{calculate_all_sample_chain_repertoire_metrics}
\usage{
calculate_all_sample_chain_repertoire_metrics(input_dataframe_list)
}
\arguments{
\item{input_dataframe_list}{a clonotype dataframe with chain name information}
}
\value{
a wide table with all samples' all chains' all repertoire metrics
}
\description{
calculate_all_sample_chain_repertoire_metrics
}
