% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_IGH_isotype_proportion.R
\name{calculate_IGH_isotype_proportion}
\alias{calculate_IGH_isotype_proportion}
\title{calculate_IGH_isotype_proportion}
\usage{
calculate_IGH_isotype_proportion(input_dataframe)
}
\arguments{
\item{input_dataframe}{a clonotype dataframe with chain name and isotype name information}
}
\value{
a dataframe of count and proportion of isotypes of IGH chain.
}
\description{
calculate_IGH_isotype_proportion
}
\examples{
calculate_IGH_isotype_proportion(convergent_clonotype_merged_dataframe)

}
