% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire_metrics_functions.R
\name{calculate_repertoire_metrics}
\alias{calculate_repertoire_metrics}
\title{calculate_repertoire_metrics}
\usage{
calculate_repertoire_metrics(named_species_vector)
}
\arguments{
\item{named_species_vector}{a named vector}
}
\value{
a named vector
}
\description{
calculate_repertoire_metrics
}
