% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_proportion_by_clones.R
\name{calibrate_proportion_by_clones}
\alias{calibrate_proportion_by_clones}
\title{calibrate_proportion_by_clones}
\usage{
calibrate_proportion_by_clones(a_dataframe)
}
\arguments{
\item{a_dataframe}{a dataframe}
}
\value{
a dataframe
}
\description{
calibrate_proportion_by_clones
}
