% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chain_separated_dataframe}
\alias{chain_separated_dataframe}
\title{a dataframe separated by chain name}
\format{
A data frame with m rows and n variables:
}
\source{
\url{https://en.wikipedia.org/wiki/T-cell_receptor}
}
\usage{
chain_separated_dataframe
}
\description{
an example dataframe separated by chain name
}
\keyword{datasets}
