% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire_metrics_functions.R
\name{combine_all_sample_repertoire_metrics}
\alias{combine_all_sample_repertoire_metrics}
\title{combine_all_sample_repertoire_metrics}
\usage{
combine_all_sample_repertoire_metrics(input_dataframe_list)
}
\arguments{
\item{input_dataframe_list}{a list of dataframes with repertoire metrics by chain name}
}
\value{
a wide table with all samples' all chains' all repertoire metrics
}
\description{
combine_all_sample_repertoire_metrics
}
