% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completely_tidy_up_clonotype.R
\name{completely_tidy_up_clonotype}
\alias{completely_tidy_up_clonotype}
\title{completely_tidy_up_clonotype}
\usage{
completely_tidy_up_clonotype(
  input_dataframe,
  adopted_clonotyping_tool = "trust"
)
}
\arguments{
\item{input_dataframe}{an original clonotype dataframe}

\item{adopted_clonotyping_tool}{a string to denote the clonotypeing tool. Default is trust}
}
\value{
a clean dataframe
}
\description{
completely_tidy_up_clonotype
}
