% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_repertoire_metrics_by_chain_name.R
\name{compute_repertoire_metrics_by_chain_name}
\alias{compute_repertoire_metrics_by_chain_name}
\title{compute_repertoire_metrics_by_chain_name}
\usage{
compute_repertoire_metrics_by_chain_name(input_dataframe)
}
\arguments{
\item{input_dataframe}{a clonotype dataframe with chain name information}
}
\value{
repertoire metrics by chain name.
}
\description{
compute_repertoire_metrics_by_chain_name
}
\examples{
compute_repertoire_metrics_by_chain_name(convergent_clonotype_merged_dataframe)

}
