% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{convergent_clonotype_merged_dataframe}
\alias{convergent_clonotype_merged_dataframe}
\title{a dataframe where convergent clonotypes are merged}
\format{
A data frame with m rows and n variables:
}
\source{
\url{https://www.frontiersin.org/articles/10.3389/fimmu.2019.02985/full}
}
\usage{
convergent_clonotype_merged_dataframe
}
\description{
an example dataframe where convergent clonotypes are merged
}
\keyword{datasets}
