% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_clonotype_to_immunarch_style.R
\name{format_clonotype_to_immunarch_style}
\alias{format_clonotype_to_immunarch_style}
\title{Convert clonotype dataframe to immunarch format}
\usage{
format_clonotype_to_immunarch_style(input_dataframe, clonotyping_tool)
}
\arguments{
\item{input_dataframe}{a clonotype dataframe from an upstream clonotyping tool}

\item{clonotyping_tool}{choose from c("mixcr, "trust")}
}
\value{
a clonotype dataframe of immunarch format
}
\description{
Convert clonotype dataframe to immunarch format
}
\examples{
format_clonotype_to_immunarch_style(raw_input_clonotype_dataframe, "trust")

}
