% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_name_x_sample_name_for_each_metric.R
\name{get_item_name_x_sample_name_for_each_metric}
\alias{get_item_name_x_sample_name_for_each_metric}
\title{get_item_name_x_sample_name_for_each_metric}
\usage{
get_item_name_x_sample_name_for_each_metric(input_dataframe)
}
\arguments{
\item{input_dataframe}{a dataframe with repertoire metrics calculated for each item name}
}
\value{
a list of dataframes that is one single repertoire metric for all items in all samples
}
\description{
get_item_name_x_sample_name_for_each_metric
}
\examples{
get_item_name_x_sample_name_for_each_metric(all_sample_all_chain_all_metrics_wide_format_dataframe)

}
