% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immunarch_style_dataframe}
\alias{immunarch_style_dataframe}
\title{An converted dataframe to immunarch format}
\format{
A data frame with m rows and n variables:
}
\source{
\url{https://github.com/immunomind/immunarch/}
}
\usage{
immunarch_style_dataframe
}
\description{
An example dataframe converted to immunarch format
}
\keyword{datasets}
