% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_convergent_clonotype.R
\name{merge_convergent_clonotype}
\alias{merge_convergent_clonotype}
\title{Merge convergent clonotype}
\usage{
merge_convergent_clonotype(input_dataframe)
}
\arguments{
\item{input_dataframe}{A dataframe of clonotype of immunarch format}
}
\value{
A dataframe of clonotype of immunarch format
merge_convergent_clonotype_function(an_immunarch_format_clonotype_dataframe)
}
\description{
Merge convergent clonotype
}
\examples{
merge_convergent_clonotype(chain_separated_dataframe)

}
