% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonproductive_CDR3aa_removed_dataframe}
\alias{nonproductive_CDR3aa_removed_dataframe}
\title{a dataframe with nonproductive CDR3aa removed}
\format{
A data frame with m rows and n variables:
}
\source{
\url{https://mixcr.readthedocs.io/en/master/analyze.html}
}
\usage{
nonproductive_CDR3aa_removed_dataframe
}
\description{
an example dataframe with nonproductive CDR3aa removed
}
\keyword{datasets}
