% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_clonotype_to_immunarch_style.R
\name{parse_immunoseq}
\alias{parse_immunoseq}
\title{Parse immunoseq format}
\usage{
parse_immunoseq(input_clone_dataframe)
}
\arguments{
\item{input_clone_dataframe}{Generated from immunoseq program}
}
\value{
A dataframe of immunarch format
}
\description{
Parse immunoseq format
}
