% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_clonotype_to_immunarch_style.R
\name{parse_mixcr}
\alias{parse_mixcr}
\title{Parse mixcr format}
\usage{
parse_mixcr(input_clone_dataframe)
}
\arguments{
\item{input_clone_dataframe}{Generated from mixcr program}
}
\value{
A dataframe of immunarch format
}
\description{
Parse mixcr format
}
