% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_clonotype_to_immunarch_style.R
\name{parse_trust4}
\alias{parse_trust4}
\title{Parse trust4 format}
\usage{
parse_trust4(input_clone_dataframe)
}
\arguments{
\item{input_clone_dataframe}{Generated from trust4 program}
}
\value{
A dataframe of immunarch format
}
\description{
Parse trust4 format
}
