% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_input_clonotype_dataframe}
\alias{raw_input_clonotype_dataframe}
\title{Raw clonotype output from trust4}
\format{
A data frame with m rows and n variables:
}
\source{
\url{https://github.com/liulab-dfci/TRUST4}
}
\usage{
raw_input_clonotype_dataframe
}
\description{
An example dataframe for raw clonotype identified by trust4
}
\keyword{datasets}
