% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_nonproductive_CDR3aa.R
\name{remove_nonproductive_CDR3aa}
\alias{remove_nonproductive_CDR3aa}
\title{Clean the clonotype dataframe}
\usage{
remove_nonproductive_CDR3aa(input_dataframe)
}
\arguments{
\item{input_dataframe}{A dataframe of clonotype of immunarch format}
}
\value{
A dataframe of clonotype of immunarch format
}
\description{
Clean the clonotype dataframe
}
\examples{
remove_nonproductive_CDR3aa(immunarch_style_dataframe)

}
