% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_convergent_clonotype.R
\name{summarize_convergence_function}
\alias{summarize_convergence_function}
\title{summarize_convergence_function}
\usage{
summarize_convergence_function(a_dataframe)
}
\arguments{
\item{a_dataframe}{a dataframe}
}
\value{
a dataframe
}
\description{
summarize_convergence_function
}
