% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_repertoire_metrics_by_chain_name.R
\name{summarize_repertoire_metrics}
\alias{summarize_repertoire_metrics}
\title{summarize_repertoire_metrics}
\usage{
summarize_repertoire_metrics(one_sample_dataframe)
}
\arguments{
\item{one_sample_dataframe}{a dataframe of clonotype in immunarch format}
}
\value{
a named vector
}
\description{
summarize_repertoire_metrics
}
