% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{the_divergent_clonotype_dataframe_list}
\alias{the_divergent_clonotype_dataframe_list}
\title{a list of dataframe where convergent clonotypes are merged}
\format{
A list of data frame with m rows and n variables:
}
\source{
\url{https://www.frontiersin.org/articles/10.3389/fimmu.2019.02985/full}
}
\usage{
the_divergent_clonotype_dataframe_list
}
\description{
an list of example dataframes where convergent clonotypes are merged
}
\keyword{datasets}
