% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_repertoire_metrics_by_chain_name.R
\name{calculate_repertoire_metrics}
\alias{calculate_repertoire_metrics}
\title{calculate_repertoire_metrics}
\usage{
calculate_repertoire_metrics(named_species_vector)
}
\arguments{
\item{named_species_vector}{a named vector}
}
\value{
a named vector
}
\description{
calculate_repertoire_metrics
}
\examples{
a_named_vector <- c("apple" = 1, "banana" = 3, "cranberry" = 7, "date" = 10)
calculate_repertoire_metrics(a_named_vector)

}
