% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_all_sample_repertoire_metrics.R
\name{combine_all_sample_repertoire_metrics}
\alias{combine_all_sample_repertoire_metrics}
\title{combine_all_sample_repertoire_metrics}
\usage{
combine_all_sample_repertoire_metrics(input_dataframe_list)
}
\arguments{
\item{input_dataframe_list}{a list of dataframes with repertoire metrics by chain name}
}
\value{
a wide table with all samples' all chains' all repertoire metrics
}
\description{
combine_all_sample_repertoire_metrics
}
\examples{
combine_all_sample_repertoire_metrics(all_sample_all_chain_all_metrics_wide_format_dataframe_list)
combine_all_sample_repertoire_metrics(all_sample_IGH_chain_all_metrics_wide_format_dataframe_list)

}
